pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "esxi-analyzer"
version = "1.2.0"
description = "Automated diagnostic tool for VMware ESXi servers - analyzes logs, detects issues, and provides troubleshooting recommendations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Raphael", email = "kekzl@users.noreply.github.com" }
]
keywords = [
    "vmware",
    "esxi",
    "diagnostics",
    "troubleshooting",
    "log-analysis",
    "virtualization",
    "sysadmin",
    "automation",
    "ssh",
    "devops"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Logging",
    "Typing :: Typed",
]
dependencies = [
    "paramiko>=3.5.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "types-paramiko>=3.5.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
esxi-analyzer = "esxi_analyzer:main"

[project.urls]
Homepage = "https://github.com/kekzl/esxi-analyzer"
Documentation = "https://github.com/kekzl/esxi-analyzer#readme"
Repository = "https://github.com/kekzl/esxi-analyzer.git"
Issues = "https://github.com/kekzl/esxi-analyzer/issues"
Changelog = "https://github.com/kekzl/esxi-analyzer/releases"

[tool.hatch.build.targets.wheel]
packages = ["lib"]

[tool.hatch.build.targets.sdist]
include = [
    "/lib",
    "/esxi_analyzer.py",
    "/config.yaml",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # ruff-specific
]
ignore = [
    "PLR0912",  # too many branches
    "PLR0913",  # too many arguments
    "PLR0915",  # too many statements
    "PLR2004",  # magic value comparison
    "PLW0603",  # global statement
    "PLC0415",  # import not at top level (needed for some test setups)
]

[tool.ruff.lint.isort]
known-first-party = ["lib"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
strict_optional = false
warn_redundant_casts = true
warn_unused_ignores = false
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "paramiko.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=lib",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["lib"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
